rename outlier_detection.utils to outlier_detection
