+++ romancal/outlier_detection/outlier_detection_step.py
@@ -4,7 +4,7 @@ import logging
 from functools import partial
 
 from romancal.datamodels.fileio import open_dataset
-from romancal.outlier_detection.utils import detect_outliers
+from romancal.outlier_detection.outlier_detection import detect_outliers
 
 from ..stpipe import RomanStep
